pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.68.0",
    "defusedxml>=0.7.1",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.7",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "tavily-python>=0.7.12",
    "uvicorn>=0.35.0",
]
